+++ app/src/main/java/com/tosslab/jandi/app/events/files/FileUploadProgressEvent.java
@@ -4,15 +4,15 @@ package com.tosslab.jandi.app.events.files;
  * Created by Steve SeongUg Jung on 15. 6. 17..
  */
 public class FileUploadProgressEvent {
-    private final int entity;
+    private final long entity;
     private final int progressPercent;
 
-    public FileUploadProgressEvent(int entity, int progressPercent) {
+    public FileUploadProgressEvent(long entity, int progressPercent) {
         this.entity = entity;
         this.progressPercent = progressPercent;
     }
 
-    public int getEntity() {
+    public long getEntity() {
         return entity;
     }
