+++ app/src/main/java/com/tosslab/jandi/app/files/upload/FilePickerViewModel.java
@@ -17,15 +17,15 @@ public interface FilePickerViewModel {
     int TYPE_UPLOAD_TAKE_PHOTO = 0x01;
     int TYPE_UPLOAD_EXPLORER = 0x02;
 
-    void selectFileSelector(int type, Fragment fragment, int entityId);
+    void selectFileSelector(int type, Fragment fragment, long entityId);
 
     void selectFileSelector(int type, Activity activity);
 
     List<String> getFilePath(Context context, int requestCode, Intent intent);
 
-    void startUpload(Activity activity, String title, int entityId, String realFilePath, String comment);
+    void startUpload(Activity activity, String title, long entityId, String realFilePath, String comment);
 
-    void showFileUploadDialog(Context context, FragmentManager fragmentManager, String realFilePath, int entityId);
+    void showFileUploadDialog(Context context, FragmentManager fragmentManager, String realFilePath, long entityId);
 
     void moveInsertFileCommnetActivity(Context context, List<String> realFilePath, int entityId);
