+++ app/src/main/java/com/tosslab/jandi/app/network/client/privatetopic/messages/GroupMessageApiV2Client.java
@@ -26,8 +26,8 @@ public interface GroupMessageApiV2Client {
     // Private Group의 Message 리스트 정보 획득
     @GET("/privateGroups/{groupId}/messages?type=old")
     @Headers("Accept:" + JandiConstants.HTTP_ACCEPT_HEADER_DEFAULT)
-    ResMessages getGroupMessages(@Query("teamId") int teamId, @Path("groupId") int groupId,
-                                 @Query("linkId") int fromId, @Query("count") int count);
+    ResMessages getGroupMessages(@Query("teamId") long teamId, @Path("groupId") long groupId,
+                                 @Query("linkId") long fromId, @Query("count") int count);
 
     @GET("/privateGroups/{groupId}/messages?type=old")
     @Headers("Accept:" + JandiConstants.HTTP_ACCEPT_HEADER_DEFAULT)
@@ -40,24 +40,24 @@ public interface GroupMessageApiV2Client {
 
     @GET("/privateGroups/{groupId}/messages?type=new")
     @Headers("Accept:" + JandiConstants.HTTP_ACCEPT_HEADER_DEFAULT)
-    ResMessages getGroupMessagesUpdatedForMarker(@Query("teamId") int teamId, @Path("groupId") int groupId,
-                                                 @Query("linkId") int currentLinkId);
+    ResMessages getGroupMessagesUpdatedForMarker(@Query("teamId") long teamId, @Path("groupId") long groupId,
+                                                 @Query("linkId") long currentLinkId);
 
     @GET("/privateGroups/{groupId}/messages?type=new")
     @Headers("Accept:" + JandiConstants.HTTP_ACCEPT_HEADER_DEFAULT)
-    ResMessages getGroupMessagesUpdatedForMarker(@Query("teamId") int teamId, @Path("groupId") int groupId,
-                                                 @Query("linkId") int currentLinkId, @Query("count") int count);
+    ResMessages getGroupMessagesUpdatedForMarker(@Query("teamId") long teamId, @Path("groupId") long groupId,
+                                                 @Query("linkId") long currentLinkId, @Query("count") int count);
 
     // Updated 된 Private Group의 리스트 정보 획득
     @GET("/privateGroups/{groupId}/messages")
     @Headers("Accept:" + JandiConstants.HTTP_ACCEPT_HEADER_DEFAULT)
-    ResMessages getGroupMarkerMessages(@Query("teamId") int teamId, @Path("groupId") int groupId,
-                                       @Query("linkId") int currentLinkId);
+    ResMessages getGroupMarkerMessages(@Query("teamId") long teamId, @Path("groupId") long groupId,
+                                       @Query("linkId") long currentLinkId);
 
     // Private Group에서의 Message 생성
     @POST("/privateGroups/{privateGroupId}/message")
     @Headers("Accept:" + JandiConstants.HTTP_ACCEPT_HEADER_V3)
-    ResCommon sendGroupMessage(@Path("privateGroupId") int privateGroupId, @Query("teamId") int teamId,
+    ResCommon sendGroupMessage(@Path("privateGroupId") long privateGroupId, @Query("teamId") long teamId,
                                @Body ReqSendMessageV3 reqSendMessageV3);
 
     // Private Group Message 수정
@@ -69,7 +69,7 @@ public interface GroupMessageApiV2Client {
     // Private Group Message 삭제
     @DELETEWithBody("/privateGroups/{groupId}/messages/{messageId}")
     @Headers("Accept:" + JandiConstants.HTTP_ACCEPT_HEADER_DEFAULT)
-    ResCommon deletePrivateGroupMessage(@Query("teamId") int teamId, @Path("groupId") int groupId,
-                                        @Path("messageId") int messageId);
+    ResCommon deletePrivateGroupMessage(@Query("teamId") long teamId, @Path("groupId") long groupId,
+                                        @Path("messageId") long messageId);
 
 }
