+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqUpdateAnnouncementStatus.java
@@ -4,15 +4,15 @@ package com.tosslab.jandi.app.network.models;
  * Created by tonyjs on 15. 6. 24..
  */
 public class ReqUpdateAnnouncementStatus {
-    private int topicId;
+    private long topicId;
     private boolean opened;
 
-    public ReqUpdateAnnouncementStatus(int topicId, boolean opened) {
+    public ReqUpdateAnnouncementStatus(long topicId, boolean opened) {
         this.topicId = topicId;
         this.opened = opened;
     }
 
-    public int getTopicId() {
+    public long getTopicId() {
         return topicId;
     }
