+++ app/src/main/java/com/tosslab/jandi/app/ui/message/v2/adapter/viewholder/CollapseStickerCommentViewHolder.java
@@ -31,7 +31,7 @@ public class CollapseStickerCommentViewHolder implements BodyViewHolder {
     }
 
     @Override
-    public void bindData(ResMessages.Link link, int teamId, int roomId, int entityId) {
+    public void bindData(ResMessages.Link link, long teamId, long roomId, long entityId) {
         ResMessages.CommentStickerMessage message = (ResMessages.CommentStickerMessage) link.message;
 
         StickerManager.getInstance().loadStickerNoOption(ivSticker, message.content.groupId, message.content.stickerId);
@@ -49,7 +49,7 @@ public class CollapseStickerCommentViewHolder implements BodyViewHolder {
     }
 
     @Override
-    public void setLastReadViewVisible(int currentLinkId, int lastReadLinkId) {
+    public void setLastReadViewVisible(long currentLinkId, long lastReadLinkId) {
         if (currentLinkId == lastReadLinkId) {
             vLastRead.setVisibility(View.VISIBLE);
         } else {
